+++ pkg/remote/trans/detection/server_handler_windows.go
@@ -1,0 +1,1 @@
+package detection
