+++ pkg/remote/trans/netpoll/listen_config.go
@@ -1,0 +1,75 @@
+//go:build !windows
+// +build !windows
+
+/*
+ * Copyright 2022 CloudWeGo Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package netpoll
+
+import (
+	"net"
+	"syscall"
+
+	"github.com/cloudwego/kitex/pkg/remote"
+	"golang.org/x/sys/unix"
+)
+
+// newListenConfig return an new net.ListenConfig.
+func newListenConfig(opt *remote.ServerOption) net.ListenConfig {
+	return net.ListenConfig{
+		Control: func(network, address string, c syscall.RawConn) error {
+			var err error
+			c.Control(func(fd uintptr) {
+				if !opt.ReusePort {
+					return
+				}
+				/* The real behavior of 'SO_REUSEADDR/SO_REUSEPORT' depends on the underlying OS.
+				For BSD(/Darwin):
+					With 'SO_REUSEADDR' option, the socket can be successfully bound unless there is
+					a conflict with another socket bound to exactly the same combination of source address and port.
+
+					Here is an example to give a better overview:
+					SO_REUSEADDR       socketA        socketB       Result
+					---------------------------------------------------------------------
+					  ON/OFF       192.168.0.1:21   192.168.0.1:21    Error (EADDRINUSE)
+					  ON/OFF       192.168.0.1:21      10.0.0.1:21    OK
+					  ON/OFF          10.0.0.1:21   192.168.0.1:21    OK
+					   OFF             0.0.0.0:21   192.168.1.0:21    Error (EADDRINUSE)
+					   OFF         192.168.1.0:21       0.0.0.0:21    Error (EADDRINUSE)
+					   ON              0.0.0.0:21   192.168.1.0:21    OK
+					   ON          192.168.1.0:21       0.0.0.0:21    OK
+					  ON/OFF           0.0.0.0:21       0.0.0.0:21    Error (EADDRINUSE)
+
+					With 'SO_REUSEPORT', you could bind an arbitrary number of sockets to exactly the same source address
+					and port as long as all prior bound sockets also had 'SO_REUSEPORT' set before they were bound.
+				For Linux < 3.9:
+					Only the option 'SO_REUSEADDR' existed.
+				For Linux >= 3.9:
+					The 'SO_REUSEPORT' behaves exactly like the option in BSD.
+
+				More details can be found at https://stackoverflow.com/questions/14388706/how-do-so-reuseaddr-and-so-reuseport-differ.
+				*/
+				if err = unix.SetsockoptInt(int(fd), unix.SOL_SOCKET, unix.SO_REUSEADDR, 1); err != nil {
+					return
+				}
+				if err = unix.SetsockoptInt(int(fd), unix.SOL_SOCKET, unix.SO_REUSEPORT, 1); err != nil {
+					return
+				}
+			})
+			return err
+		},
+	}
+}
