+++ pkg/remote/trans/netpoll/trans_server.go
@@ -30,11 +30,9 @@ import (
 	"syscall"
 
 	"github.com/cloudwego/kitex/pkg/klog"
-	"github.com/cloudwego/netpoll"
-	"golang.org/x/sys/unix"
-
 	"github.com/cloudwego/kitex/pkg/remote"
 	"github.com/cloudwego/kitex/pkg/utils"
+	"github.com/cloudwego/netpoll"
 )
 
 // NewTransServerFactory creates a default netpoll transport server factory.
@@ -48,47 +46,7 @@ type netpollTransServerFactory struct{}
 func (f *netpollTransServerFactory) NewTransServer(opt *remote.ServerOption, transHdlr remote.ServerTransHandler) remote.TransServer {
 	return &transServer{
 		opt: opt, transHdlr: transHdlr,
-		lncfg: net.ListenConfig{Control: func(network, address string, c syscall.RawConn) error {
-			var err error
-			c.Control(func(fd uintptr) {
-				if !opt.ReusePort {
-					return
-				}
-				/* The real behavior of 'SO_REUSEADDR/SO_REUSEPORT' depends on the underlying OS.
-				For BSD(/Darwin):
-					With 'SO_REUSEADDR' option, the socket can be successfully bound unless there is
-					a conflict with another socket bound to exactly the same combination of source address and port.
-
-					Here is an example to give a better overview:
-					SO_REUSEADDR       socketA        socketB       Result
-					---------------------------------------------------------------------
-					  ON/OFF       192.168.0.1:21   192.168.0.1:21    Error (EADDRINUSE)
-					  ON/OFF       192.168.0.1:21      10.0.0.1:21    OK
-					  ON/OFF          10.0.0.1:21   192.168.0.1:21    OK
-					   OFF             0.0.0.0:21   192.168.1.0:21    Error (EADDRINUSE)
-					   OFF         192.168.1.0:21       0.0.0.0:21    Error (EADDRINUSE)
-					   ON              0.0.0.0:21   192.168.1.0:21    OK
-					   ON          192.168.1.0:21       0.0.0.0:21    OK
-					  ON/OFF           0.0.0.0:21       0.0.0.0:21    Error (EADDRINUSE)
-
-					With 'SO_REUSEPORT', you could bind an arbitrary number of sockets to exactly the same source address
-					and port as long as all prior bound sockets also had 'SO_REUSEPORT' set before they were bound.
-				For Linux < 3.9:
-					Only the option 'SO_REUSEADDR' existed.
-				For Linux >= 3.9:
-					The 'SO_REUSEPORT' behaves exactly like the option in BSD.
-
-				More details can be found at https://stackoverflow.com/questions/14388706/how-do-so-reuseaddr-and-so-reuseport-differ.
-				*/
-				if err = unix.SetsockoptInt(int(fd), unix.SOL_SOCKET, unix.SO_REUSEADDR, 1); err != nil {
-					return
-				}
-				if err = unix.SetsockoptInt(int(fd), unix.SOL_SOCKET, unix.SO_REUSEPORT, 1); err != nil {
-					return
-				}
-			})
-			return err
-		}},
+		lncfg: newListenConfig(opt),
 	}
 }
